Fixed problem with "clean" shutdown failing due to crash and active log set to "clean".
git-svn-id: https://svn.neo4j.org/components/kernel/trunk@4264 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ advanced/src/main/java/org/neo4j/kernel/impl/nioneo/store/NeoStore.java
@@ -119,6 +119,11 @@ public class NeoStore extends AbstractStore
 
     public void flushAll()
     {
+        if ( relTypeStore == null || propStore == null || relStore == null || 
+                nodeStore == null )
+        {
+            return;
+        }
         relTypeStore.flushAll();
         propStore.flushAll();
         relStore.flushAll();
